+++ store/types/gas_test.go
@@ -16,10 +16,13 @@ func TestInfiniteGasMeter(t *testing.T) {
 	meter.ConsumeGas(10, "consume 10")
 	require.Equal(t, uint64(10), meter.GasConsumed())
 	require.Equal(t, uint64(10), meter.GasConsumedToLimit())
+	meter.RefundGas(1, "refund 1")
+	require.Equal(t, uint64(9), meter.GasConsumed())
 	require.False(t, meter.IsPastLimit())
 	require.False(t, meter.IsOutOfGas())
 	meter.ConsumeGas(Gas(math.MaxUint64/2), "consume half max uint64")
 	require.Panics(t, func() { meter.ConsumeGas(Gas(math.MaxUint64/2)+2, "panic") })
+	require.Panics(t, func() { meter.RefundGas(meter.GasConsumed()+1, "refund greater than consumed") })
 }
 
 func TestGasMeter(t *testing.T) {
@@ -57,6 +60,11 @@ func TestGasMeter(t *testing.T) {
 		require.Panics(t, func() { meter.ConsumeGas(1, "") }, "Exceeded but not panicked. tc #%d", tcnum)
 		require.Equal(t, meter.GasConsumedToLimit(), meter.Limit(), "Gas consumption (to limit) not match limit")
 		require.Equal(t, meter.GasConsumed(), meter.Limit()+1, "Gas consumption not match limit+1")
+
+		require.NotPanics(t, func() { meter.RefundGas(1, "refund 1") })
+		require.Equal(t, meter.GasConsumed(), meter.Limit(), "Gas consumption not match limit+1")
+		require.Panics(t, func() { meter.RefundGas(meter.GasConsumed()+1, "refund greater than consumed") })
+
 		meter2 := NewGasMeter(math.MaxUint64)
 		meter2.ConsumeGas(Gas(math.MaxUint64/2), "consume half max uint64")
 		require.Panics(t, func() { meter2.ConsumeGas(Gas(math.MaxUint64/2)+2, "panic") })